Add AsyncCommand<T> with strongly typed parameter
In the WPF MVVM world, it is fairly common for projects to write their
own simple command implementations, sometimes called DelegateCommand,
RelayCommand or simply Command. Since ICommand uses object as the
parameter type, it is also common for developers using CommandParameter
often to implement a generic Command<T> to avoid casting boilerplate.

For the same reason, this adds a strongly typed AsyncCommand<T> that
simply wraps an AsyncCommand, relaying CanExecutedChanged and
PropertyChanged events as well as the Execute call, casting the
parameter.